feat(frontend): Address review comments part 1
