+++ frontend/src/globals.css
@@ -67,8 +67,9 @@
   --light-dark-background: #130713;
   --text-gray: #798795;
   --modal-border: #170f2e;
-  --warning-colour:rgba(255, 251, 0, 0.84);
-  --warning-colour-alt: rgb(70, 70, 1);
+  --warning-colour:#BDC000;
+  --warning-colour-alt: #272A0A;
+  --warning-text-colour: #F0F0F0;
 }
 
 body {
