--- frontend/src/hooks/useScrollTracker.js
@@ -1,27 +1,0 @@
-import { useRef, useState, useEffect } from "react";
-
-const useScrollTracker = () => {
-  const scrollRef = useRef(null);
-  const [activeIndex, setActiveIndex] = useState(0);
-
-  const handleScroll = () => {
-    if (scrollRef.current) {
-      const { scrollLeft, scrollWidth, clientWidth } = scrollRef.current;
-      const scrollPercentage = scrollLeft / (scrollWidth - clientWidth);
-      const index = Math.round(scrollPercentage);
-      setActiveIndex(index);
-    }
-  };
-
-  useEffect(() => {
-    const scrollContainer = scrollRef.current;
-    if (scrollContainer) {
-      scrollContainer.addEventListener("scroll", handleScroll);
-      return () => scrollContainer.removeEventListener("scroll", handleScroll);
-    }
-  }, []);
-
-  return { scrollRef, activeIndex, setActiveIndex };
-};
-
-export default useScrollTracker;
