+++ frontend/src/pages/forms/Form.jsx
@@ -3,7 +3,7 @@ import { ReactComponent as ETH } from '../../assets/icons/ethereum.svg';
 import { ReactComponent as USDC } from '../../assets/icons/borrow_usdc.svg';
 import { ReactComponent as STRK } from '../../assets/icons/strk.svg';
 import TokenSelector from 'components/TokenSelector';
-import BalanceCards from 'components/BalanceCards';
+import BalanceCards from 'components/BalanceCards/BalanceCards';
 import MultiplierSelector from 'components/MultiplierSelector';
 import { handleTransaction } from 'services/transaction';
 import Spinner from 'components/spinner/Spinner';
