+++ frontend/src/pages/spotnet/dashboard/withdraw-all/withdraw_all.css
@@ -1,12 +1,3 @@
-body {
-    font-family:
-      'Rethink Sans',
-      -apple-system,
-      Roboto,
-      Helvetica,
-      sans-serif;
-  }
-  
   .withdrawall-wrapper {
     position: relative;
     background: url('../../../../../public/Form-bg.png') no-repeat;
